refactor : notice item 컴포넌트 위치
